fix(CVPage): move CVPage component to pages directory and update import path
