+++ src/App.tsx
@@ -3,9 +3,9 @@ import { useEffect } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 import Portfolio from "./pages/Portfolio";
+import CVPage from "./pages/CvPage";
 
 import Lenis from "lenis";
-import CVPage from "./sections/CvPage";
 
 const App = () => {
   // Lenis Logic
